pyproject.toml
[tool.poetry]
name = "funnel-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyrogram = "^2.0.106"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
tgcrypto = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
